// src/type/enum.type.ts
export enum Lang {
  en_US = 'en-US',
  zh_TW = 'zh-TW',
}

export enum PageType {
  home,
  page,
  undefined,
}

export enum Device {
  laptop,
  mobile,
}

export enum DesignType {
  foundation = 'foundation',
  development = 'development',
}

export enum UICategory {
  // ---- Foundation
  theme = 'theme',
  layout = 'layout',
  typography = 'typography',
  symbol = 'symbol',
  elevation = 'elevation',
  // ---- Development
  navigationBar = 'navigationBar',
  tabs = 'tabs',
  buttons = 'buttons',
  sidebar = 'sidebar',
  ratingBar = 'ratingBar',
  hero = 'hero',
  breadcrumb = 'breadcrumb',
  alerts = 'alerts',
  timeline = 'timeline',
  tags = 'tags',
  progressBar = 'progressBar',
  avatar = 'avatar',
  badge = 'badge',
  footer = 'footer',
  skeleton = 'skeleton',
  spinner = 'spinner',
  stepper = 'stepper',
  list = 'list',
  cards = 'cards',
  forms = 'forms',
  modal = 'modal',
  dropdown = 'dropdown',
  disclosure = 'disclosure',
  popover = 'popover',
  carousel = 'carousel',
  tooltip = 'tooltip',
  pagination = 'pagination',
  toast = 'toast',
}

export enum HeaderAction {
  youtube = 'Youtube',
  translate = 'translate',
  theme = 'themeMode',
}

export enum NavLinkType {
  tab,
  item,
}

export enum Level {
  first,
  second,
  third,
}

export enum Install {
  cdn = 'CDN',
  npm = 'NPM',
}

export enum ColorAnatomy {
  primary = 'Primary',
  secondary = 'Secondary',
  neutral = 'Neutral',
  accent = 'Accent',
}

export enum IconCategory {
  application = 'Application',
  unit = 'Unit',
  weather = 'Weather',
}

export enum Icon {
  // ---- application
  mic = 'mic',
  share = 'share',
  setting = 'setting',
  edit = 'edit',
  update = 'update',
  trash = 'trash',
  calendar = 'calendar',
  search = 'search',
  start = 'start',
  timer = 'timer',
  lock = 'lock',
  unlock = 'unlock',
  scan = 'scan',
  loading = 'loading',
  grid = 'grid',
  sensor = 'sensor',
  camera = 'camera',
  person = 'person',
  home = 'home',
  download = 'download',
  battery = 'battery',
  done = 'done',
  visible = 'visible',
  hide = 'hide',
  video = 'video',
  phone = 'phone',
  chart = 'chart',
  attach = 'attach',
  filter = 'filter',
  cart = 'cart',
  plus = 'plus',
  subtract = 'subtract',
  notification = 'notification',
  arrow = 'arrow',
  internet = 'internet',
  file = 'file',
  printer = 'printer',
  zoomin = 'zoomin',
  zoomout = 'zoomout',
  help = 'help',
  heart = 'heart',
  more = 'more',
  email = 'email',
  services = 'services',
  wifi = 'wifi',
  location = 'location',
  image = 'image',
  mouse = 'mouse',
  hamburger = 'hamburger',
  // ---- unit
  humi = 'humi',
  temp = 'temp',
  heat = 'heat',
  percent = 'percent',
  potenz = 'potenz',
  pm10 = 'pm10',
  'pm2.5' = 'pm2.5',
  carbon = 'carbon',
  // ---- weather
  morning = 'morning',
  sunrise = 'sunrise',
  nuit = 'nuit',
  cloudy = 'cloudy',
  rain = 'rain',
  overcast = 'overcast',
  squall = 'squall',
  torrential = 'torrential',
  thunder = 'thunder',
  drizzle = 'drizzle',
  snow = 'snow',
}

export enum CodeType {
  jsx = 'JSX',
  html = 'HTML',
}